+++ Pontu/include/model/Piece.h
@@ -10,6 +10,7 @@
 #include "Island.h"
 //#include "Player.h" //If we use a pointer for the player instead of an int 
 #include <stdbool.h>
+#include <stddef.h>
 
 /**
  *\Struct Piece
@@ -18,7 +19,7 @@
  */
 typedef struct 
 {
-	int idJ; ///< Player's id (owner of the piece)
+	size_t idJ; ///< Player's id (owner of the piece)
 	bool stuck; ///< Evaluate to true if the piece is isolated
 	Island island; ///< Piece's localisation on the game board
 }Piece;
@@ -27,6 +28,6 @@ typedef struct
  * \brief Creates a new Piece
  * \param[in]	idJ	The ID (index in array) of the Player owning the Piece
  */
-Piece newPiece(const int idJ);
+Piece newPiece(const size_t idJ);
 
 #endif // PIECE_H
