+++ Pontu/src/model/Board.c
@@ -1,12 +1,12 @@
 #include "model/Board.h"
 
-Board newBoard(const int nbPlayers) {
-	const int nbPiecesPerPlayer = (nbPlayers == 4) ? 2 : 3;
+Board newBoard(const size_t nbPlayers) {
+	const int nbPiecesPerPlayer = (nbPlayers == 4u) ? 2 : 3;
 	Board board = {.nbPieces = nbPiecesPerPlayer*nbPlayers};
 	int pieceIndex = 0;
 
 	// Init pieces for each player
-	for (int player_i = 0; player_i < nbPlayers; player_i++) {
+	for (size_t player_i = 0; player_i < nbPlayers; player_i++) {
 		for (int piece_i = 0; piece_i < nbPiecesPerPlayer; piece_i++) {
 			board.arrPieces[pieceIndex] = newPiece(player_i);
 			pieceIndex++;
